added router.get find by email
