+++ client/src/App.js
@@ -5,7 +5,6 @@ import Dashboard from './pages/Dashboard';
 import Login from './pages/Login';
 import Directory from './pages/Directory';
 import AddUser from './pages/AddUser';
-import Contact from './pages/Contact';
 import './style.css';
 
 function App() {
@@ -17,7 +16,6 @@ function App() {
       <Route exact path='/Dashboard' component={Dashboard}/>
       <Route exact path='/Directory' component={Directory}/>
       <Route exact path='/AddUser' component={AddUser}/>
-      <Route exact path='/Contact' component={Contact}/>
       </div>
     </Router>
