+++ models/employee.js
@@ -7,7 +7,7 @@ if (process.env.JAWSDB_URL) {
   });
 } else {
 
-   sequelize = new Sequelize("sequelize_library", "root", "amillionwords", {
+   sequelize = new Sequelize("sequelize_library", "root", "22087005552493", {
 
     host: "localhost",
     port: 3306,
